changed name of function, export/import validation functions, moved to seperate file becaause of ecmascript processing of helpers.js file
