+++ packages/minimongo/minimongo.js
@@ -1,3 +1,5 @@
+import { assertHasValidFieldNames } from './validation.js';
+
 // XXX type checking on selectors (graceful error if malformed)
 
 // LocalCollection: a set of documents that supports queries and modifiers.
@@ -540,15 +542,13 @@ LocalCollection.Cursor.prototype._depend = function (changers, _allow_unordered)
   }
 };
 
-// XXX enforce rule that field names can't start with '$' or contain '.'
-// (real mongodb does in fact enforce this)
 // XXX possibly enforce that 'undefined' does not appear (we assume
 // this in our handling of null and $exists)
 LocalCollection.prototype.insert = function (doc, callback) {
   var self = this;
   doc = EJSON.clone(doc);
 
-  hasValidFieldNames(doc);
+  assertHasValidFieldNames(doc);
 
   if (!_.has(doc, '_id')) {
     // if you really want to use ObjectIDs, set this global.
